rename files to camelCasing
