+++ src/mode/modeInsert.ts
