+++ src/mode/modeNormal.ts
